headers: Properly guard includes and fix headers
Since the refactoring in naming conventions and the move to a different
source directory, this library has stagnated and not been updated in a
while. This is mainly due to its lack of inclusion in the build process
of the main kernel. Simply fix this now and once the kernel is in a
state of development when such a library is required, integrate
accordingly.

Signed-off-by: Aun-Ali Zaidi <admin@kodeit.net>